fix: PK long -> int

+++ src/main/java/com/doo/skeleton/repository/first/example/MariaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.doo.skeleton.repository.first.example.vo.MariaDto;
 
-public interface MariaRepository extends JpaRepository<MariaDto, Long> {
+public interface MariaRepository extends JpaRepository<MariaDto, Integer> {
 
 }
